flask/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
async-timeout==4.0.3
attrs==23.1.0
blis==0.7.11
catalogue==2.0.10
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
cloudpathlib==0.15.1
confection==0.1.3
cymem==2.0.8
dataclasses-json==0.5.14
dnspython==2.4.2
frozenlist==1.4.0
idna==3.4
Jinja2==3.1.2
langchain==0.0.295
langcodes==3.3.0
langsmith==0.0.38
loguru==0.7.2
MarkupSafe==2.1.3
marshmallow==3.20.1
multidict==6.0.4
murmurhash==1.0.10
mypy-extensions==1.0.0
numexpr==2.8.6
numpy==1.26.0
openai==0.28.1
packaging==23.1
pathy==0.10.2
pinecone-client==2.2.4
preshed==3.0.9
pydantic==2.3.0
pydantic_core==2.6.3
pypdf==3.16.1
PyPDF2==3.0.1
python-dateutil==2.8.2
PyYAML==6.0.1
regex==2023.10.3
requests==2.31.0
six==1.16.0
smart-open==6.4.0
spacy==3.7.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.21
srsly==2.4.8
tenacity==8.2.3
thinc==8.2.1
tiktoken==0.5.1
tqdm==4.66.1
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.8.0
urllib3==2.0.4
wasabi==1.1.2
weasel==0.3.2
yarl==1.9.2
